Refactored GET /pets: moved filtering, sorting and pagination logic into PetService for cleaner separation
